pyproject.toml
[project]
name = "argus"
version = "0.1.0"
description = "Nuclear power plant foreign object visual detection system"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.10.0",
    "ultralytics>=8.3.0",
    "anomalib>=2.0.0,<3.0.0",
    "openvino>=2024.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "scikit-image>=0.24.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    "websockets>=13.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
argus = "argus.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/argus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
